🚚 rename: Bookmark 도메인 분리
