+++ http/test.http
@@ -67,7 +67,7 @@ Authorization: Bearer {{matsterToken}}
 
 ### 북마크 등록 및 취소
 POST http://localhost:8080/api/v1/me/announcement/0/bookmarked
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZW1haWwiOiJzcG9udXNAZ21haWwuY29tIiwiYXV0aCI6IlNUVURFTlQiLCJpYXQiOjE3MDY5NTcwNTAsImV4cCI6MTcwNzk1NzA1MH0.y-RlHw9kL08-3ca5JRAOimtzB8rNPhlL9oIqUH6rFOY
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZW1haWwiOiJzcG9udXNAZ21haWwuY29tIiwiYXV0aCI6IlNUVURFTlQiLCJpYXQiOjE3MDY5NTg4MDcsImV4cCI6MTcwNzk1ODgwN30.TmFCrtJteVt_2rCRij3oofD6VHdXardFlYlOn3y7y7Q
 Content-Type: application/json
 
 {
