+++ src/main/java/com/sponus/sponusbe/domain/bookmark/controller/BookmarkController.java
@@ -1,16 +1,15 @@
-package com.sponus.sponusbe.domain.organization.controller;
+package com.sponus.sponusbe.domain.bookmark.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.sponus.sponusbe.auth.annotation.AuthOrganization;
-import com.sponus.sponusbe.domain.organization.dto.BookmarkToggleRequest;
-import com.sponus.sponusbe.domain.organization.dto.BookmarkToggleResponse;
+import com.sponus.sponusbe.domain.bookmark.dto.BookmarkToggleRequest;
+import com.sponus.sponusbe.domain.bookmark.dto.BookmarkToggleResponse;
 import com.sponus.sponusbe.domain.organization.entity.Organization;
-import com.sponus.sponusbe.domain.organization.service.BookmarkService;
+import com.sponus.sponusbe.domain.bookmark.service.BookmarkService;
 import com.sponus.sponusbe.global.common.ApiResponse;
 
 import lombok.RequiredArgsConstructor;
