+++ src/main/java/com/sponus/sponusbe/domain/bookmark/dto/BookmarkToggleRequest.java
@@ -1,7 +1,7 @@
-package com.sponus.sponusbe.domain.organization.dto;
+package com.sponus.sponusbe.domain.bookmark.dto;
 
 import com.sponus.sponusbe.domain.announcement.entity.Announcement;
-import com.sponus.sponusbe.domain.organization.entity.Bookmark;
+import com.sponus.sponusbe.domain.bookmark.entity.Bookmark;
 import com.sponus.sponusbe.domain.organization.entity.Organization;
 
 public record BookmarkToggleRequest(
