+++ src/main/java/com/sponus/sponusbe/domain/bookmark/repository/BookmarkRepository.java
@@ -1,11 +1,11 @@
-package com.sponus.sponusbe.domain.organization.repository;
+package com.sponus.sponusbe.domain.bookmark.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sponus.sponusbe.domain.announcement.entity.Announcement;
-import com.sponus.sponusbe.domain.organization.entity.Bookmark;
+import com.sponus.sponusbe.domain.bookmark.entity.Bookmark;
 import com.sponus.sponusbe.domain.organization.entity.Organization;
 
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
